Redo driver, log, and test infrastructure
Driver exists in a separate crate `crates/driver`, allowing it to be
easily reused between the main program and testing.

Logging has replaced usage of the `log` crate with a more rudimentary
implementation. This impl also writes all logs to a buffer in debug
mode, which can be used by the test harness.

Additional fixes:
  - Grammar for identifier patterns accidentally consumed pattern twice
  - "atomic expression" instead of "atomic pattern"