Start new MIR implementation
Create MIR with datatype that is ANF by construction, along with join points.
This fixes many issues with the previous ANF conversion that incorrectly handled ifs.
The new conversion uses a defunctionalized CPS conversion. The algorithm introduces
unnecessary administrative redexes/let bindings. I could change the algorithm to be
smarter, like something in "No-Brainer CPS Conversion", or rely on a shrinking pass
after in order to keep the conversion simpler.